Make build_modssl_with_npn.sh automatically resumable at any point
r=bmcquade

+++ src/build_modssl_with_npn.sh
@@ -31,8 +31,8 @@
 MODSSL_SO_DESTPATH=$(pwd)/mod_ssl.so
 
 if [ -f $MODSSL_SO_DESTPATH ]; then
-echo "mod_ssl already exists at $MODSSL_SO_DESTPATH. Please remove."
-exit 1
+  echo "mod_ssl already exists at $MODSSL_SO_DESTPATH. Please remove."
+  exit 1
 fi
 
 if [ -z "$BUILDROOT" ]; then
@@ -43,36 +43,57 @@ else
 fi
 
 if [ ! -d "$BUILDROOT" ]; then
-echo "Not a directory: $BUILDROOT"
-exit 1
+  echo "Not a directory: $BUILDROOT"
+  exit 1
 fi
 
-function do_cleanup {
+# Convert BUILDROOT to an absolute path.
+BUILDROOT="$(cd $(dirname $BUILDROOT); pwd)/$(basename $BUILDROOT)"
+echo "Using buildroot: $BUILDROOT"
 echo ""
-echo "Build aborted."
-if [ $REMOVE_BUILDROOT -eq 1 ]; then
-  echo -n "Cleaning up ... "
-  rm -rf "$BUILDROOT"
-  echo "done"
-fi
-exit 1
+
+function do_cleanup {
+  echo ""
+  echo "Build aborted."
+  if [ $REMOVE_BUILDROOT -eq 1 ]; then
+    echo -n "Cleaning up ... "
+    rm -rf "$BUILDROOT"
+    echo "done"
+  fi
+  exit 1
 }
 
 trap 'do_cleanup' SIGINT SIGTERM
 
-function download_file {
-echo "Downloading $1"
-curl -# "$1" -o $2 || do_cleanup
-if [[ $(md5sum $2 | cut -d\  -f1) != $3 ]]; then
-  echo "md5sum mismatch for $2"
+PROGRESS_DIR=$BUILDROOT/progress
+mkdir -p $PROGRESS_DIR
+if [ $? -ne 0 ]; then
   do_cleanup
 fi
+
+function download_file {
+  if [ ! -f "$PROGRESS_DIR/$2.downloaded" ]; then
+    echo "Downloading $1"
+    curl -# "$1" -o $2 || do_cleanup
+    if [[ $(md5sum $2 | cut -d\  -f1) != $3 ]]; then
+      echo "md5sum mismatch for $2"
+      do_cleanup
+    fi
+    touch "$PROGRESS_DIR/$2.downloaded"
+  else
+    echo "Already downloaded $1"
+  fi
 }
 
 function uncompress_file {
-echo -n "Uncompressing $1 ... "
-tar xzf $1 || do_cleanup
-echo "done"
+  if [ ! -f "$PROGRESS_DIR/$1.uncompressed" ]; then
+    echo -n "Uncompressing $1 ... "
+    tar xzf $1 || do_cleanup
+    echo "done"
+    touch "$PROGRESS_DIR/$1.uncompressed"
+  else
+    echo "Already uncompressed $1"
+  fi
 }
 
 OPENSSL_SRC_TGZ_URL="http://www.openssl.org/source/openssl-1.0.1-beta2.tar.gz"
@@ -91,10 +112,6 @@ APACHE_HTTPD_BUILDLOG=$(mktemp -p /tmp httpd_buildlog.XXXXXXXXXX)
 
 pushd $BUILDROOT >/dev/null
 
-# If a build fails part way through you can set RESUMEBUILD=1 to skip
-# the download/unpack/configure steps.
-if [[ $RESUMEBUILD -ne 1 ]]; then
-
 download_file $OPENSSL_SRC_TGZ_URL $OPENSSL_SRC_TGZ ef66ad92539014e1a8fe33bdd8159bad
 download_file $APACHE_HTTPD_SRC_TGZ_URL $APACHE_HTTPD_SRC_TGZ b24ca6db942a4f8e57c357e5e3058d31
 download_file $APACHE_HTTPD_MODSSL_NPN_PATCH_URL $APACHE_HTTPD_MODSSL_NPN_PATCH cfd98e0b0b13f86825df7610b2437e5a
@@ -104,71 +121,96 @@ echo ""
 uncompress_file $OPENSSL_SRC_TGZ
 uncompress_file $APACHE_HTTPD_SRC_TGZ
 
-pushd $APACHE_HTTPD_SRC_ROOT >/dev/null
-echo -n "Applying Apache mod_ssl NPN patch ... "
-patch -p0 <$BUILDROOT/$APACHE_HTTPD_MODSSL_NPN_PATCH
-if [ $? -ne 0 ]; then
-echo "Failed to patch."
-do_cleanup
+if [ ! -f "$PROGRESS_DIR/modssl_patched" ]; then
+  pushd $APACHE_HTTPD_SRC_ROOT >/dev/null
+  echo -n "Applying Apache mod_ssl NPN patch ... "
+  patch -p0 < $BUILDROOT/$APACHE_HTTPD_MODSSL_NPN_PATCH
+  if [ $? -ne 0 ]; then
+    echo "Failed to patch."
+    do_cleanup
+  fi
+  echo "done"
+  popd >/dev/null  # $APACHE_HTTPD_SRC_ROOT
+  touch "$PROGRESS_DIR/modssl_patched"
+else
+  echo "Already applied Apache mod_ssl NPN patch."
 fi
-echo "done"
-popd >/dev/null
 
 echo ""
 
-pushd $OPENSSL_SRC_ROOT >/dev/null
-echo -n "Configuring OpenSSL ... "
-./config no-shared -fPIC >> $OPENSSL_BUILDLOG
-if [ $? -ne 0 ]; then
-echo "Failed. Build log at $OPENSSL_BUILDLOG."
-do_cleanup
+if [ ! -f "$PROGRESS_DIR/openssl_configured" ]; then
+  pushd $OPENSSL_SRC_ROOT >/dev/null
+  echo -n "Configuring OpenSSL ... "
+  ./config no-shared -fPIC >> $OPENSSL_BUILDLOG
+  if [ $? -ne 0 ]; then
+    echo "Failed. Build log at $OPENSSL_BUILDLOG."
+    do_cleanup
+  fi
+  echo "done"
+  popd >/dev/null  # $OPENSSL_SRC_ROOT
+  touch "$PROGRESS_DIR/openssl_configured"
+else
+  echo "Already configured OpenSSL."
 fi
-echo "done"
-popd >/dev/null
 
-fi  # [[ $RESUMEBUILD -ne 1 ]]
-
-pushd $OPENSSL_SRC_ROOT >/dev/null
-echo -n "Building OpenSSL (this may take a while) ... "
-make INSTALL_PREFIX=$(pwd) install >> $OPENSSL_BUILDLOG 2>&1
-if [ $? -ne 0 ]; then
-echo "Failed. Build log at $OPENSSL_BUILDLOG."
-do_cleanup
+if [ ! -f "$PROGRESS_DIR/openssl_built" ]; then
+  pushd $OPENSSL_SRC_ROOT >/dev/null
+  echo -n "Building OpenSSL (this may take a while) ... "
+  make INSTALL_PREFIX=$(pwd) install >> $OPENSSL_BUILDLOG 2>&1
+  if [ $? -ne 0 ]; then
+    echo "Failed. Build log at $OPENSSL_BUILDLOG."
+    do_cleanup
+  fi
+  echo "done"
+  popd >/dev/null  # $OPENSSL_SRC_ROOT
+  touch "$PROGRESS_DIR/openssl_built"
+else
+  echo "Already built OpenSSL."
 fi
-echo "done"
-popd >/dev/null
 
 rm -f "$OPENSSL_BUILDLOG"
 
 echo ""
 
-pushd $APACHE_HTTPD_SRC_ROOT >/dev/null
-echo -n "Configuring Apache mod_ssl ... "
-./configure --enable-ssl=shared --with-ssl=$BUILDROOT/$OPENSSL_SRC_ROOT/usr/local/ssl >> $APACHE_HTTPD_BUILDLOG
-if [ $? -ne 0 ]; then
-echo "Failed. Build log at $APACHE_HTTPD_BUILDLOG."
-do_cleanup
+if [ ! -f "$PROGRESS_DIR/modssl_configured" ]; then
+  pushd $APACHE_HTTPD_SRC_ROOT >/dev/null
+  echo -n "Configuring Apache mod_ssl ... "
+  ./configure --enable-ssl=shared --with-ssl=$BUILDROOT/$OPENSSL_SRC_ROOT/usr/local/ssl >> $APACHE_HTTPD_BUILDLOG
+  if [ $? -ne 0 ]; then
+    echo "Failed. Build log at $APACHE_HTTPD_BUILDLOG."
+    do_cleanup
+  fi
+  echo "done"
+  popd >/dev/null  # $APACHE_HTTPD_SRC_ROOT
+  touch "$PROGRESS_DIR/modssl_configured"
+else
+  echo "Already configured Apache mod_ssl."
 fi
-echo "done"
 
-echo -n "Building Apache mod_ssl (this may take a while) ... "
-make >> $APACHE_HTTPD_BUILDLOG 2>&1
-if [ $? -ne 0 ]; then
-echo "Failed. Build log at $APACHE_HTTPD_BUILDLOG."
-do_cleanup
+if [ ! -f "$PROGRESS_DIR/modssl_built" ]; then
+  pushd $APACHE_HTTPD_SRC_ROOT >/dev/null
+  echo -n "Building Apache mod_ssl (this may take a while) ... "
+  make >> $APACHE_HTTPD_BUILDLOG 2>&1
+  if [ $? -ne 0 ]; then
+    echo "Failed. Build log at $APACHE_HTTPD_BUILDLOG."
+    do_cleanup
+  fi
+  echo "done"
+  popd >/dev/null  # $APACHE_HTTPD_SRC_ROOT
+  touch "$PROGRESS_DIR/modssl_built"
+else
+  echo "Already built Apache mod_ssl."
 fi
-echo "done"
-popd >/dev/null
 
 rm -f "$APACHE_HTTPD_BUILDLOG"
 
-popd >/dev/null
+popd >/dev/null  # $BUILDROOT
 
 MODSSL_SO_SRCPATH=$(find $BUILDROOT/$APACHE_HTTPD_SRC_ROOT -name mod_ssl.so)
 if [ $(echo $MODSSL_SO_SRCPATH | wc -l) -ne 1 ]; then
-echo "Found multiple mod_ssl.so's:"
-echo $MODSSL_SO_SRCPATH
-do_cleanup
+  echo "Found multiple mod_ssl.so's:"
+  echo $MODSSL_SO_SRCPATH
+  do_cleanup
 fi
 
 cp $MODSSL_SO_SRCPATH $MODSSL_SO_DESTPATH
